refactor: merge bool and string code back to a single file
